+++ packages/orq-rc/src/models/operations/createprompt.ts
@@ -1353,7 +1353,6 @@ export const CreatePromptProvider = {
   Nvidia: "nvidia",
   Jina: "jina",
   Elevenlabs: "elevenlabs",
-  Litellm: "litellm",
   Cerebras: "cerebras",
   Openailike: "openailike",
   Bytedance: "bytedance",