+++ packages/orq-rc/src/models/operations/getallprompts.ts
@@ -421,7 +421,6 @@ export const GetAllPromptsProvider = {
   Nvidia: "nvidia",
   Jina: "jina",
   Elevenlabs: "elevenlabs",
-  Litellm: "litellm",
   Cerebras: "cerebras",
   Openailike: "openailike",
   Bytedance: "bytedance",