+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-19T06:47:32.188Z",
+    "2026-06-19T08:56:30.928Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {